+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.4.4] - 2021-05-06
 ### Changed
 - Updating trace.py classes to accept draw keywords.
+
+### Added
+- `render_text` function to render text; matplotlib now required.
 
 ## [0.4.3] - 2021-04-28
 ### Fixed